Fixed slow mode to use Date.getTime() and updated to v1.0.4

+++ src/ben_dude56/plugins/bencmd/chat/ChatCommands.java
@@ -232,7 +232,7 @@ public class ChatCommands implements Commands {
 							"You used a blocked word..."));
 			return;
 		}
-		int slowTimeLeft = plugin.chatListen.slow.playerBlocked(user.getName());
+		long slowTimeLeft = plugin.chatListen.slow.playerBlocked(user.getName());
 		if ((!plugin.perm.userFile.hasPermission(user.getName(),
 				"ignoreSlowMode", true, true))
 				&& plugin.chatListen.slow.isEnabled()) {
@@ -293,7 +293,7 @@ public class ChatCommands implements Commands {
 							"You used a blocked word..."));
 			return;
 		}
-		int slowTimeLeft = plugin.chatListen.slow.playerBlocked(user.getName());
+		long slowTimeLeft = plugin.chatListen.slow.playerBlocked(user.getName());
 		if ((!plugin.perm.userFile.hasPermission(user.getName(),
 				"ignoreSlowMode", true, true))
 				&& plugin.chatListen.slow.isEnabled()) {
